# o2-sim: Simple z-dependent R-cut (for ZDC)