--- a/Common/SimConfig/include/SimConfig/SimCutParams.h
+++ b/Common/SimConfig/include/SimConfig/SimCutParams.h
@@ -26,8 +26,11 @@ struct SimCutParams : public o2::conf::ConfigurableParamHelper<SimCutParams> {
 
   double maxRTracking = 1E20;    // max R tracking cut in cm (in the VMC sense) -- applied in addition to cutting in the stepping function
   double maxAbsZTracking = 1E20; // max |Z| tracking cut in cm (in the VMC sense) -- applied in addition to cutting in the stepping function
-  double ZmaxA = 1E20;           // max Z tracking cut on A side in cm -- applied in the stepping function
-  double ZmaxC = 1E20;           // max Z tracking cut on C side in cm -- applied in the stepping function
+  double ZmaxA = 1E20;           // max Z tracking cut on A side in cm -- applied in our custom stepping function
+  double ZmaxC = 1E20;           // max Z tracking cut on C side in cm -- applied in out custom stepping function
+
+  float maxRTrackingZDC = 50; // R-cut applied in the tunnel leading to ZDC when z > beampipeZ (custom stepping function)
+  float tunnelZ = 1900;       // Z-value from where we apply maxRTrackingZDC (default value taken from standard "hall" dimensions)
 
   O2ParamDef(SimCutParams, "SimCutParams");
 };