Implementation of a simple z-position dependent
R-cut. This is needed since VMC only offers
global R-cuts.

Here we apply a much tigher R-cut in the
tunnel leading to ZDC, to avoid excessive
stepping in the concrete. This is configurable via SimCutParams.