@davidrohr 2021-03-26
Hi, for reference, I did some benchmarks to compare the performance and the output (1 identical Pb-Pb event in all cases):
- Old version, default settings, without ZDC: 83s CPU time
- New version, default settings, without ZDC: 82s CPU time
- Old version, with ZDC, maxRTracking = 50cm: 51s CPU time
- New version, default settings, with ZDC: 85s CPU time
- Old version, default settings, with ZDC: 90s CPU time