@davidrohr 2021-03-26
(also ran the same settings twice to check if the results are consistent, and it always agrees within <1sec.)
Judging only from the o2sim_HitsZDC size, this is identical for all the version with ZDC included, so it seems we are not loosing something.
For all other detectors it is also mostly identical, except for the version with maxRTracking = 50cm obviously.
I think the new version is generally superior, since it shows the same performance as before without ZDC and better performance with ZDC.
The only thing that I am wondering a bit: Some time ago I believe to remember that ZDC prolonged the simulation by almost 2x, while now even in the old version it is only a 10% effect. Perhaps there have been also other optimizations in between?